fix: vlQuantidadeSaco to big decimal

+++ src/main/java/com/corretora/api/dto/contrato/ContratoDto.java
@@ -34,7 +34,7 @@ public class ContratoDto {
 
     private BigDecimal vlComissao;
 
-    private Integer vlQuantidadeSaco;
+    private BigDecimal vlQuantidadeSaco;
 
     private String dsPadraoTolerancia;
